rebease for pipe util - pipe util changes for common and spark
