+++ common/core/src/main/java/zingg/util/DFWriter.java
@@ -1,0 +1,11 @@
+package zingg.util;
+
+public interface DFWriter<D,R,C> {
+
+    public void setMode(String s);
+    public DFWriter<D,R,C> format(String f);
+    public DFWriter<D,R,C> option(String k, String v);
+    public void save(String location);
+    public void save();
+    
+}
