+++ common/core/src/main/java/zingg/util/PipeUtilBase.java
@@ -38,11 +38,19 @@ public interface PipeUtilBase<S, D, R, C> {
 
 	public Pipe<D, R, C> getStopWordsPipe(Arguments args, String string);
 
-	public String getPipesAsString(Pipe<D, R, C>[] pipes);
+	public String getPipesAsString(Pipe<D,R,C>[] pipes);
 
 	public S getSession();
 
 	public void setSession(S session);
 
+	public Pipe<D,R,C> setOverwriteMode(Pipe<D,R,C> p);
+
+	public ZFrame<D,R,C> getZFrame(ZFrame<D,R,C> z);
+
+	public ZFrame<D,R,C> addLineNo (ZFrame<D,R,C> input);
+
+	public DFWriter<D,R,C> getWriter(ZFrame<D,R,C> input);
+	public DFReader<D,R,C> getReader();
 	
 }
